refactor: 객체 구조 수정
- 패키지 분리
- 패키지 이름 수정
- MovingStrategy 클래스 작성으로 더 비즈니스적인 이름으로 변경
