+++ src/main/java/car_racing2/domain/MovingStrategy.java
@@ -1,0 +1,5 @@
+package car_racing2.domain;
+
+public interface MovingStrategy {
+    boolean isMovable();
+}
