+++ src/main/java/car_racing2/domain/model/RaceResult.java
@@ -1,4 +1,4 @@
-package car_racing2.model;
+package car_racing2.domain.model;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,23 +7,22 @@ public class RaceResult {
     private static final String CAR_RESULT_DELIMITER = "\n";
     private static final String WINNERS_DELIMITER = ", ";
 
-    private List<String> resultOfRounds = new ArrayList<>();
-    private String winners = "";
+    private final List<String> resultOfRounds = new ArrayList<>();
+    private final StringBuilder winners = new StringBuilder();
 
     public void recordResultOfRound(List<String> carsNameWithPosition) {
         this.resultOfRounds.add(String.join("", carsNameWithPosition));
     }
 
     public void recordRaceWinners(List<String> winners) {
-        this.winners = String.join(WINNERS_DELIMITER, winners);
+        this.winners.append(String.join(WINNERS_DELIMITER, winners));
     }
 
     public String getResultOfRounds() {
-        String result = String.join(CAR_RESULT_DELIMITER, resultOfRounds);
-        return result;
+        return String.join(CAR_RESULT_DELIMITER, resultOfRounds);
     }
 
     public String getWinners() {
-        return winners;
+        return winners.toString();
     }
 }
